+++ src/test/data/object/formatTree.js
@@ -1,0 +1,48 @@
+import formatTree from './../../../data/object/formatTree'
+
+(function() {
+  // formatTree相关
+  const list = [
+    {
+      pid: 3,
+      id: 6,
+      name: '3-6'
+    },
+    {
+      pid: 0,
+      id: 2,
+      name: '2'
+    },
+    {
+      pid: 0,
+      id: 1,
+      name: '1'
+    },
+    {
+      pid: 2,
+      id: 3,
+      name: '2-3'
+    },
+    {
+      pid: 1,
+      id: 4,
+      name: '1-4'
+    },
+    {
+      pid: 2,
+      id: 5,
+      name: '2-5'
+    }
+  ]
+  let treeList = formatTree(list, {
+    parentId: 'pid'
+  })
+  if (treeList.length != 2 || treeList[0].name != '2' || treeList[1].name != '1') {
+    console.error('formatTree未成功,长度或者顺序错误')
+  } else {
+    let insideData = treeList[0].children[0].children[0]
+    if (!insideData || insideData.name != '3-6') {
+      console.error('formatTree未成功，内部值错误')
+    }
+  }
+})();