+++ src/test/index.js
@@ -10,6 +10,7 @@ function loadContents(contents, fn) {
 const contents = require.context('./data', true, /\.js$/)
 if (getEnv('real') == 'development') {
   loadContents(contents, function(item) {
-    let data = item.default || item
+    // let data = item.default || item
+    // console.log(item, data)
   })
 }